+++ config/constants.py
@@ -1,4 +1,4 @@
-APP_VERSION = "v2.0.7"
+APP_VERSION = "v2.0.9"
 GITHUB_REPO = "netbox-community/devicetype-library"
 BRANCH = "master"
 RULES_FILE = "naming_rules.json"